// src/infrastructure/openProject/wpStatus.enum.ts
enum WPStatus {
  new = "New",
  confirmed = "Confirmed",
  inSpecification = "In specification",
  specified = "Specified",
  inProgress = "In progress",
  developed = "Developed",
  inTesting = "In testing",
  tested = "Tested",
  testFailed = "Test failed",
  onHold = "On hold",
  closed = "Closed",
  rejected = "Rejected",
}
export default WPStatus;